+++ tests/testthat/disable_logging_output.R
@@ -4,6 +4,9 @@ library(tryCatchLog)
 # test helper script to disable the package-internal logging output
 # (as futile.logger's "flog.threshold("FATAL")" does)
 
+# Note: Always use the FQN for this function in unit tests since
+#       otherwise strange side effects may occur (logging is still
+#       enabled or disabled - perhaps due to stacked testthat environments).
 tryCatchLog::set.logging.functions(
     error.log.func = function(msg) invisible()
   , warn.log.func  = function(msg) invisible()
@@ -12,3 +15,4 @@ tryCatchLog::set.logging.functions(
   , trace.log.func = function(msg) invisible()
   , fatal.log.func = function(msg) invisible()
 )
+